+++ sdk-conference/src/main/kotlin/com/pexip/sdk/conference/ReferException.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 Pexip AS
+ * Copyright 2023 Pexip AS
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,20 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.pexip.sdk.sample.pinrequirement
+package com.pexip.sdk.conference
 
-import com.pexip.sdk.api.infinity.InfinityService
-import com.pexip.sdk.conference.Conference
-
-sealed interface PinRequirementOutput {
-
-    @JvmInline
-    value class None(val conference: Conference) : PinRequirementOutput
-
-    data class Some(
-        val step: InfinityService.ConferenceStep,
-        val required: Boolean,
-    ) : PinRequirementOutput
-
-    data class Error(val t: Throwable) : PinRequirementOutput
-}
+/**
+ * Thrown to indicate that call transfer has failed.
+ *
+ * @property cause a cause of this exception
+ */
+public class ReferException @JvmOverloads constructor(cause: Throwable? = null) :
+    RuntimeException("Failed to transfer the call.", cause)